+++ assets/sensorData/sensorData/src/firebase/firebase.h
@@ -2,6 +2,6 @@
 #define FIREBASE_H
 
 void setupFirebase();
-void sendDataToFirebase(double tempValue, int soundValue, int smokeValue, double fuelLvlValue, int vibrationValue, int currentValue, int oilPressureValue);
+void sendDataToFirebase(double tempValue, int soundValue, int smokeValue, double fuelLvlValue, int vibrationValue, double currentValue, int oilPressureValue);
 
 #endif
